refactor: API 요청 생성 로직 분리
