+++ src/constants/constant.ts
@@ -1,11 +1,19 @@
 export const BASE_URL = {
+  TMDB_BASE_URL: "https://api.themoviedb.org/3",
   POSTER_BASE_URL: "https://image.tmdb.org/t/p/w200",
   HERO_BASE_URL: "https://image.tmdb.org/t/p/w1920_and_h800_multi_faces",
 };
+
+export const DEFAULT_LANGUAGE = "ko-KR";
 
 export const SKELETON_MOVIE_COUNT = 20;
 
 export const IMAGE_URL = {
   STAR_IMAGE_URL: "/images/star_empty.png",
   DEFAULT_THUMBNAIL_IMAGE_URL: `/images/default-thumbnail.jpeg`,
+};
+
+export const API_PATH = {
+  POPULAR_MOVIE: `/movie/popular`,
+  SEARCH_MOVIE: `/search/movie`,
 };
